requirements.txt
django
django-filter
django-ckeditor
python-dotenv